LayoutTests/imported/w3c/web-platform-tests/dom/observable/tentative/observable-reduce.any-expected.txt


PASS reduce(): Reduces the values of the Observable, starting with the initial seed value
PASS reduce(): Rejects if the source observable emits an error
PASS reduce(): Seeds with the first value of the source, if no initial value is provided
PASS reduce(): Errors thrown in reducer reject the promise and abort the source
PASS reduce(): When source is empty, promise resolves with initial value
PASS reduce(): When source is empty, AND no seed value is provided, the promise rejects with a TypeError
PASS reduce(): Reject with an AbortError if the subscription is aborted before the source completes
PASS reduce(): Reduces the values for different objects
